add dedicated error types for block comment errors
