+++ parser/comment_test.go
@@ -19,15 +19,221 @@
 package parser
 
 import (
+	"math/big"
 	"testing"
 
+	"github.com/onflow/cadence/ast"
+	"github.com/onflow/cadence/parser/lexer"
+	. "github.com/onflow/cadence/test_utils/common_utils"
 	"github.com/stretchr/testify/require"
 )
 
-func TestParseBlockCommentEmpty(t *testing.T) {
+func TestParseBlockComment(t *testing.T) {
 
 	t.Parallel()
 
-	_, err := testParseProgram(`/**/`)
-	require.NoError(t, err)
+	t.Run("empty", func(t *testing.T) {
+		t.Parallel()
+
+		_, errs := testParseExpression(`/**/ true`)
+		require.Empty(t, errs)
+	})
+
+	t.Run("nested", func(t *testing.T) {
+
+		t.Parallel()
+
+		result, errs := testParseExpression(" /* test  foo/* bar  */ asd*/ true")
+		require.Empty(t, errs)
+
+		AssertEqualWithDiff(t,
+			&ast.BoolExpression{
+				Value: true,
+				Range: ast.Range{
+					StartPos: ast.Position{Line: 1, Column: 30, Offset: 30},
+					EndPos:   ast.Position{Line: 1, Column: 33, Offset: 33},
+				},
+			},
+			result,
+		)
+	})
+
+	t.Run("two comments", func(t *testing.T) {
+
+		t.Parallel()
+
+		result, errs := testParseExpression(" /*test  foo*/ /* bar  */ true")
+		require.Empty(t, errs)
+
+		AssertEqualWithDiff(t,
+			&ast.BoolExpression{
+				Value: true,
+				Range: ast.Range{
+					StartPos: ast.Position{Line: 1, Column: 26, Offset: 26},
+					EndPos:   ast.Position{Line: 1, Column: 29, Offset: 29},
+				},
+			},
+			result,
+		)
+	})
+
+	t.Run("in infix", func(t *testing.T) {
+
+		t.Parallel()
+
+		result, errs := testParseExpression(" 1/*test  foo*/+/* bar  */ 2  ")
+		require.Empty(t, errs)
+
+		AssertEqualWithDiff(t,
+			&ast.BinaryExpression{
+				Operation: ast.OperationPlus,
+				Left: &ast.IntegerExpression{
+					PositiveLiteral: []byte("1"),
+					Value:           big.NewInt(1),
+					Base:            10,
+					Range: ast.Range{
+						StartPos: ast.Position{Line: 1, Column: 1, Offset: 1},
+						EndPos:   ast.Position{Line: 1, Column: 1, Offset: 1},
+					},
+				},
+				Right: &ast.IntegerExpression{
+					PositiveLiteral: []byte("2"),
+					Value:           big.NewInt(2),
+					Base:            10,
+					Range: ast.Range{
+						StartPos: ast.Position{Line: 1, Column: 27, Offset: 27},
+						EndPos:   ast.Position{Line: 1, Column: 27, Offset: 27},
+					},
+				},
+			},
+			result,
+		)
+	})
+
+	t.Run("nested, extra closing", func(t *testing.T) {
+
+		t.Parallel()
+
+		_, errs := testParseExpression(" /* test  foo/* bar  */ asd*/ true */ bar")
+		AssertEqualWithDiff(t,
+			[]error{
+				// `true */ bar` is parsed as infix operation of path
+				&SyntaxError{
+					Message: "expected token identifier",
+					Pos: ast.Position{
+						Offset: 37,
+						Line:   1,
+						Column: 37,
+					},
+					Secondary:     "check for missing punctuation, operators, or syntax elements",
+					Documentation: "https://cadence-lang.org/docs/language/syntax",
+				},
+			},
+			errs,
+		)
+	})
+
+	t.Run("missing closing", func(t *testing.T) {
+		t.Parallel()
+
+		_, errs := testParseExpression(`/*`)
+
+		AssertEqualWithDiff(t,
+			[]error{
+				&MissingCommentEndError{
+					Pos: ast.Position{Offset: 2, Line: 1, Column: 2},
+				},
+				UnexpectedEOFError{
+					Pos: ast.Position{Offset: 2, Line: 1, Column: 2},
+				},
+			},
+			errs,
+		)
+	})
+
+	t.Run("nested, missing closing", func(t *testing.T) {
+
+		t.Parallel()
+
+		_, errs := testParseExpression(" /* test  foo/* bar  */ asd true ")
+		AssertEqualWithDiff(t,
+			[]error{
+				&MissingCommentEndError{
+					Pos: ast.Position{Offset: 33, Line: 1, Column: 33},
+				},
+				UnexpectedEOFError{
+					Pos: ast.Position{Offset: 33, Line: 1, Column: 33},
+				},
+			},
+			errs,
+		)
+	})
+
+	t.Run("invalid content", func(t *testing.T) {
+
+		t.Parallel()
+
+		// The lexer should never produce such an invalid token stream in the first place
+
+		tokens := &testTokenStream{
+			tokens: []lexer.Token{
+				{
+					Type: lexer.TokenBlockCommentStart,
+					Range: ast.Range{
+						StartPos: ast.Position{
+							Line:   1,
+							Offset: 0,
+							Column: 0,
+						},
+						EndPos: ast.Position{
+							Line:   1,
+							Offset: 1,
+							Column: 1,
+						},
+					},
+				},
+				{
+					Type: lexer.TokenIdentifier,
+					Range: ast.Range{
+						StartPos: ast.Position{
+							Line:   1,
+							Offset: 2,
+							Column: 2,
+						},
+						EndPos: ast.Position{
+							Line:   1,
+							Offset: 4,
+							Column: 4,
+						},
+					},
+				},
+				{Type: lexer.TokenEOF},
+			},
+			input: []byte(`/*foo`),
+		}
+
+		_, errs := ParseTokenStream(
+			nil,
+			tokens,
+			func(p *parser) (ast.Expression, error) {
+				return parseExpression(p, lowestBindingPower)
+			},
+			Config{},
+		)
+		AssertEqualWithDiff(t,
+			[]error{
+				&UnexpectedTokenInBlockCommentError{
+					GotToken: lexer.Token{
+						SpaceOrError: nil,
+						Range: ast.Range{
+							StartPos: ast.Position{Offset: 2, Line: 1, Column: 2},
+							EndPos:   ast.Position{Offset: 4, Line: 1, Column: 4},
+						},
+						Type: lexer.TokenIdentifier,
+					},
+				},
+			},
+			errs,
+		)
+	})
 }
